made the suggested PR changes

+++ server/src/routes/api/v1/activityReviewRouter.js
@@ -1,14 +1,16 @@
 import express from "express"
-import { Review } from "../../../models/index.js"
+import { Activity, Review, User } from "../../../models/index.js"
 import { ValidationError } from "objection"
 import cleanUserInput from "../../../services/cleanUserInput.js"
 
 const activityReviewRouter = new express.Router({ mergeParams: true })
 
+
 activityReviewRouter.post("/", async (req, res) => {
+    const userId = req.user.id
     const newReview = cleanUserInput(req.body)
     const { content } = newReview
-    const userId = req.user.id
+
     const activityId = req.params.id
 
     try {
@@ -22,10 +24,34 @@ activityReviewRouter.post("/", async (req, res) => {
     }
 })
 
+
+activityReviewRouter.get("/", async (req, res) => {
+
+    const userId = req.user.id
+    const activityId = req.params.id
+
+    try {
+        const activity = await Activity.query().findById(activityId)
+        const allReviews = await activity.$relatedQuery("reviews").where('userId', userId)
+        return res.status(200).json({ reviews: allReviews })
+    } catch (error) {
+        return res.status(500).json({ errors: error })
+    }
+})
+
+
 activityReviewRouter.delete("/:id", async (req, res) => {
+
     try {
         const reviewToDelete = await Review.query().findById(req.params.id)
-        await reviewToDelete.$query().delete()
+        if (reviewToDelete.userId === req.user.id) {
+            await reviewToDelete.$query().delete()
+        } else {
+            const errorMessage = `You are not authorized to delete this review.`
+            const error = new Error(errorMessage)
+            throw error
+        }
+
         return res.status(200).json({ review: reviewToDelete })
     } catch (err) {
         return res.status(500).json({ errors: err })
